testing the limit line for sleep start and end

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/activities/charts/SleepUtils.java
@@ -22,7 +22,7 @@ public class SleepUtils {
     public static final float Y_VALUE_DEEP_SLEEP = 0.01f;
     public static final float Y_VALUE_LIGHT_SLEEP = 0.016f;
 
-    public static boolean isSleep(byte type) {
+    public static boolean isSleep(int type) {
         return type == ActivityKind.TYPE_DEEP_SLEEP || type == ActivityKind.TYPE_LIGHT_SLEEP;
     }
 }
